@@ -1,3 +1,20 @@
+buildscript {
+    repositories {
+        mavenCentral()
+        google()
+    }
+    dependencies {
+        classpath 'com.android.tools.build:gradle:8.2.0'
+    }
+}
+
+allprojects {
+    repositories {
+        mavenCentral()
+        google()
+    }
+}
+
 apply plugin: 'com.android.application'
 
 android {
@@ -5,11 +22,11 @@ android {
         abortOnError false
     }
 
-    compileSdkVersion 28
+    compileSdkVersion 33
     defaultConfig {
         applicationId "firstvan.converter"
-        minSdkVersion 28
-        targetSdkVersion 28
+        minSdkVersion 33
+        targetSdkVersion 33
         versionCode 2
         versionName "2_0"
         multiDexEnabled true
@@ -22,12 +39,14 @@ android {
         }
     }
     compileOptions {
-        targetCompatibility 1.8
-        sourceCompatibility 1.8
+        targetCompatibility 19
+        sourceCompatibility 19
     }
+
+    namespace 'com.firstvan.app'
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.+'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
 }
